% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-names.R
\name{clean_names}
\alias{clean_names}
\title{Clean names}
\usage{
clean_names(
  in_path,
  file_name,
  lname_col,
  fname_col,
  sex_col,
  other_cols,
  out_path,
  middle
)
}
\arguments{
\item{in_path}{Path to the directory where the input file is located}

\item{file_name}{Input filename}

\item{lname_col}{Raw last name column in the input file}

\item{fname_col}{Raw first name column in the input file}

\item{sex_col}{Sex column in the input file (1=male, 2=female)}

\item{other_cols}{Other columns to be kept from the input file}

\item{out_path}{Path to the directory where the output file will be saved}

\item{middle}{Middle initial column in the input file}
}
\value{
\preformatted{      NULL
}
}
\description{
This function cleans raw first and last names by removing titles (e.g., Dr.)
and replacing nicknames with standard names (e.g., Billy to William).
It creates a new dataset with cleaned first names, middle initials, and last names.
}
