% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-records.R
\name{match_records}
\alias{match_records}
\title{Match records}
\usage{
match_records(
  file_A,
  fname_var_A,
  lname_var_A,
  time_var_A,
  id_var_A,
  vars_to_keep_A,
  file_B,
  fname_var_B,
  lname_var_B,
  time_var_B,
  id_var_B,
  vars_to_keep_B,
  out_path,
  out_file_name,
  timeband = 2,
  uniqueband_file = 2,
  uniqueband_match = 2
)
}
\arguments{
\item{file_A}{Dataset A}

\item{fname_var_A}{First name column in dataset A}

\item{lname_var_A}{Last name column in dataset A}

\item{time_var_A}{Time column in dataset A}

\item{id_var_A}{ID column in dataset A}

\item{vars_to_keep_A}{Other columns to be kept from dataset A}

\item{file_B}{Dataset B}

\item{fname_var_B}{First name column in dataset B}

\item{lname_var_B}{Last name column in dataset B}

\item{time_var_B}{Time column in dataset B}

\item{id_var_B}{ID column in dataset B}

\item{vars_to_keep_B}{Other columns to be kept from dataset B}

\item{out_path}{Path to the directory where the output file will be saved}

\item{out_file_name}{Output filename}

\item{timeband}{Time band used when searching for nonexact matches}

\item{uniqueband_file}{Uniqueness band used for conservative ABE (within)}

\item{uniqueband_match}{Uniqueness band used for conservative ABE (between)}
}
\value{
\preformatted{          NULL
}
}
\description{
This function matches records in dataset A to records in dataset B
using the ABE method. The matched dataset includes both standard
and conservative ABE matches.
}
